backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Firebase
firebase-admin==6.2.0

# Database
pyrebase4==4.7.1

# Document Processing
PyPDF2==3.0.1
docx2txt==0.8
python-docx==1.1.0
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow>=10.4.0

# Text Processing & NLP
nltk==3.8.1
textstat==0.7.3
spacy==3.7.2
transformers==4.35.2
torch>=2.2.0

# Text-to-Speech
google-cloud-texttospeech==2.14.2
boto3==1.33.13  # For Amazon Polly
pyttsx3==2.90
gTTS==2.4.0

# Image Processing & AI
opencv-python==4.8.1.78
google-cloud-vision==3.5.0

# Math Processing
sympy==1.12
latex2mathml==3.76.0

# URL Processing
requests>=2.28.0,<2.30
beautifulsoup4==4.12.2
newspaper3k==0.2.8
html2text==2020.1.16

# Utilities
werkzeug==3.0.1
gunicorn==21.2.0
validators==0.22.0
